requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.29
alembic==1.13.1
python-dotenv==1.0.1
aiosqlite
greenlet
nest_asyncio